openssl

+++ 3rd/postgresql/gcc_x64.sh
@@ -4,13 +4,13 @@ source ../../make_envars.sh
 
 LIB_NAME=postgresql
 LIB_SRC=postgresql-12.1
-PREFIX=$ROOT_3RD
+PREFIX=$ROOT_3RD/$LIB_NAME
 
 echo "===  lib_src: $LIB_SRC ==== prefix: === $PREFIX ====================="
 
-#[ -n "$ROOT_DIR" ] && rm -rf $ROOT_DIR/3rd/$LIB_NAME/$LIB_SRC
+[ -n "$ROOT_DIR" ] && rm -rf $ROOT_DIR/3rd/$LIB_NAME/$LIB_SRC
 
-#$ROOT_DIR/tools/rhel/x64/7z.sh x $LIB_SRC.7z
+$ROOT_DIR/tools/rhel/x64/7z.sh x $LIB_SRC.7z
 
 cd $LIB_SRC
 
@@ -25,13 +25,13 @@ PATH="$ROOT_3RD/libxml2/bin:$ROOT_3RD/libxslt/bin;$PATH"
 #LD_LIBRARY_PATH="$ROOT_3RD/lib:$LD_LIBRARY_PATH"
 #LDFLAGS="-L$ROOT_3RD/lib"
 
-rm config.log
+[ -f config.log ] && rm config.log
 
-#autoconf -f
+autoconf -f
 sh ./configure \
-  CPPFLAGS="-I$ROOT_3RD/libxml2/include/libxml2 -I$ROOT_3RD/libxslt/include -I$ROOT_3RD/uuid/include" \
-  CFLAGS="-I$ROOT_3RD/libxml2/include/libxml2 -I$ROOT_3RD/libxslt/include -I$ROOT_3RD/uuid/include" \
-  LDFLAGS="-L$ROOT_3RD/libxml2/lib -L$ROOT_3RD/libxslt/lib -L$ROOT_3RD/uuid/lib" \
+  CPPFLAGS="-I$ROOT_3RD/libxml2/include/libxml2 -I$ROOT_3RD/libxslt/include -I$ROOT_3RD/uuid/include -I$ROOT_3RD/openssl/include" \
+  CFLAGS="-I$ROOT_3RD/libxml2/include/libxml2 -I$ROOT_3RD/libxslt/include -I$ROOT_3RD/uuid/include -I$ROOT_3RD/openssl/include" \
+  LDFLAGS="-L$ROOT_3RD/libxml2/lib -L$ROOT_3RD/libxslt/lib -L$ROOT_3RD/uuid/lib -L$ROOT_3RD/openssl/lib" \
   --enable-debug \
   --prefix=$ROOT_3RD \
   --with-openssl \
@@ -40,11 +40,6 @@ sh ./configure \
   --with-libxslt \
   --without-readline
 
-            
-#sh ./autogen.sh
-#perl Configure \
-#    --prefix=$PREFIX linux-x86_64
-
-#make install
+make install
 cd ..
